int datatype update

+++ src/trusted/service_runtime/nacl_syscall_strace.c
@@ -52,14 +52,14 @@ void NaClStraceClose(int cageid, int d, int ret) {
     fprintf(tracingOutputFile, "%d close(%d) = %d\n", cageid, d, ret);
 }
 
-void NaClStraceRead(int cageid, int d, void *buf, size_t count, int ret) {
+void NaClStraceRead(int cageid, int d, void *buf, int  count, int ret) {
     fprintf(tracingOutputFile, "%d read(%d, %p, %zu) = %d\n", cageid, d, buf, count, ret);
 }
 void NaClStraceExit(int cageid, int status){
     fprintf(tracingOutputFile, "%d exit() = %d\n", cageid, status);
 
 }
-void NaClStraceThreadExit(int cageid, int32_t *stack_flag, uint32_t zero){
+void NaClStraceThreadExit(int cageid, int *stack_flag, uint32_t zero){
     fprintf(tracingOutputFile, "%d thread_exit(%d, %u) = void\n", cageid, stack_flag ? *stack_flag : 0, zero);
 }
 
@@ -75,27 +75,27 @@ void NaClStraceDup3(int cageid, int oldfd,int newfd,int flags,int ret){
     fprintf(tracingOutputFile, "%d dup3(%d, %d, %d) = %d\n", cageid, oldfd, newfd, flags, ret);
 
 }
-void NaClStraceGetdents(int cageid, int d, void *dirp, size_t count, size_t retval, ssize_t getdents_ret, uintptr_t sysaddr) {
+void NaClStraceGetdents(int cageid, int d, void *dirp, int  count, int  retval, int  getdents_ret, uintptr_t sysaddr) {
     fprintf(tracingOutputFile, "%d getdents(%d, %p, %zu, %zu, %zd, %zu) = void\n", 
             cageid, d, dirp, count, retval, getdents_ret, sysaddr);
 }
 
 
-void NaClStracePread(int cageid, int d, void *buf, size_t count, off_t offset, int32_t retval) {
+void NaClStracePread(int cageid, int d, void *buf, int  count, off_t offset,int retval) {
     fprintf(tracingOutputFile, "%d pread(%d, %p, %zu, %lld) = %d\n", cageid, d, buf, count, (long long)offset, retval);
 }
 
 
-void NaClStraceWrite(int cageid, int d, void *buf, size_t count, int32_t ret) {
+void NaClStraceWrite(int cageid, int d, void *buf, int  count,int ret) {
     fprintf(tracingOutputFile, "%d write(%d, %p, %zu) = %d\n", cageid, d, buf, count, ret);
 }
 
-void NaClStracePWrite(int cageid, int d, const void *buf, size_t count, off_t offset, int32_t retval) {
+void NaClStracePWrite(int cageid, int d, const void *buf, int  count, off_t offset, int retval) {
     fprintf(tracingOutputFile, "%d pwrite(%d, %p, %zu, %lld) = %d\n", cageid, d, buf, count, (long long)offset, retval);
 }
 
 
-void NaClStraceLseek(int cageid, int d, int whence, uintptr_t offset, int32_t ret) {
+void NaClStraceLseek(int cageid, int d, int whence, uintptr_t offset, int ret) {
     fprintf(tracingOutputFile, "%d lseek(%d, %d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, d, whence, offset, ret);
 }
 
@@ -107,33 +107,33 @@ void NaClStraceIoctl(int cageid, int d, unsigned long request, void *arg_ptr, in
 // void NaClStraceFstat(int cageid, int d, uintptr_t result, int retval) {
 //     fprintf(tracingOutputFile, "%d fstat(%d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, d, result, retval);
 // }
-void NaClStraceFstat(int cageid, int d, const struct lind_stat *result, int32_t retval) {
+void NaClStraceFstat(int cageid, int d, const struct lind_stat *result, int retval) {
     fprintf(tracingOutputFile, "%d fstat(%d, %p) = %d\n", cageid, d, (void *)result, retval);
 }
 
-void NaClStraceStat(int cageid, char* path, uintptr_t result, int32_t retval) {
+void NaClStraceStat(int cageid, char* path, uintptr_t result, int retval) {
     fprintf(tracingOutputFile, "%d stat(%s, 0x%08"NACL_PRIxPTR") = %d\n", cageid, path, result, retval);
 }
 
 
-void NaClStraceLStat(int cageid, char* path, uintptr_t result, int32_t retval) {
+void NaClStraceLStat(int cageid, char* path, uintptr_t result, int retval) {
     fprintf(tracingOutputFile, "%d lstat(%s, 0x%08"NACL_PRIxPTR") = %d\n", cageid, path, result, retval);
 }
 
-void NaClStraceMkdir(int cageid, char* path, int mode, int32_t retval) {
+void NaClStraceMkdir(int cageid, char* path, int mode, int retval) {
     fprintf(tracingOutputFile, "%d mkdir(%s, %d) = %d\n", cageid, path, mode, retval);
 }
 
 
-void NaClStraceRmdir(int cageid, uint32_t pathname, const char* path, int32_t retval) {
+void NaClStraceRmdir(int cageid, uint32_t pathname, const char* path,int retval) {
     fprintf(tracingOutputFile, "%d rmdir(%u, %s) = %d\n", cageid, pathname, path, retval);
 }
-void NaClStraceChdir(int cageid, uint32_t pathname, const char* path, int32_t retval) {
+void NaClStraceChdir(int cageid, uint32_t pathname, const char* path,int retval) {
     fprintf(tracingOutputFile, "%d chdir(%u, %s) = %d\n", cageid, pathname, path, retval);
 }
 
 
-void NaClStraceChmod(int cageid, const char *path, int mode, int32_t retval) {
+void NaClStraceChmod(int cageid, const char *path, int mode,int retval) {
     fprintf(tracingOutputFile, "%d chmod(%s, %d) = %d\n", cageid, path, mode, retval);
 }
 
@@ -153,7 +153,7 @@ void NaClStraceFsync(int cageid, int fd, int ret) {
 void NaClStraceFdatasync(int cageid, int fd, int ret) {
     fprintf(tracingOutputFile, "%d fdatasync(%d) = %d\n", cageid, fd, ret);
 }
-void NaClStraceGetcwd(int cageid, char* buf, size_t size, uintptr_t sysaddr, int32_t retval) {
+void NaClStraceGetcwd(int cageid, char* buf, int  size, uintptr_t sysaddr,int retval) {
     fprintf(tracingOutputFile, "%d getcwd(%p, %zu, 0x%08"NACL_PRIxPTR") = %d\n",cageid, buf, size, sysaddr, retval);
 }
 
@@ -162,19 +162,19 @@ void NaClStraceLink(int cageid, char* from,char* to) {
     fprintf(tracingOutputFile, "%d link(%s, %s) = void\n", cageid, from, to);
 
 }
-void NaClStraceUnlink(int cageid, char* pathname,int32_t retval){
+void NaClStraceUnlink(int cageid, char* pathname,int  retval){
     fprintf(tracingOutputFile, "%d unlink(%s) = %d\n", cageid, pathname, retval);
 
 }
 
-void NaClStraceRename(int cageid, const char *oldpath, const char *newpath, int32_t retval) {
+void NaClStraceRename(int cageid, const char *oldpath, const char *newpath,int retval) {
     fprintf(tracingOutputFile, "%d rename(oldpath: \"%s\", newpath: \"%s\") = %d\n", cageid, oldpath, newpath, retval);
 }
-void NaClStraceMmap(int cageid, void *start,size_t length,int prot,int flags,int d, uintptr_t offset, int32_t retval) {
+void NaClStraceMmap(int cageid, void *start,int  length,int prot,int flags,int d, uintptr_t offset,int retval) {
     fprintf(tracingOutputFile, "%d mmap(%p, %zu, %d, %d, %d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, start, length, prot, flags, d, offset, retval);
 
 }
-// void NaClStraceMunmap(int cageid, uintptr_t sysaddr, size_t length,int32_t retval) {
+// void NaClStraceMunmap(int cageid, uintptr_t sysaddr, int  length,int32_t retval) {
 //    fprintf(tracingOutputFile, "%d munmap(0x%08"NACL_PRIxPTR", %zu) = %d\n", cageid, sysaddr, length, retval);
 
 // }
@@ -182,82 +182,82 @@ void NaClStraceMmap(int cageid, void *start,size_t length,int prot,int flags,int
 void NaClStraceShmat(int cageid, int shmid, void *shmaddr, int shmflg, int retval) {
     fprintf(tracingOutputFile, "%d shmat(%d, %p, %d) = %d\n", cageid, shmid, shmaddr, shmflg, retval);
 }
-void NaClStraceShmget(int cageid, int key, size_t size, int shmflg, int retval) {
+void NaClStraceShmget(int cageid, int key, int  size, int shmflg, int retval) {
     fprintf(tracingOutputFile, "%d shmget(%d, %zu, %d) = %d\n", cageid, key, size, shmflg, retval);
 }
 void NaClStraceShmdt(int cageid, void *shmaddr, int retval) {
     fprintf(tracingOutputFile, "%d shmdt(%p) = %d\n", cageid, shmaddr, retval);
 }
 
-void NaClStraceShmctl(int cageid, int shmid, int cmd, int32_t retval) {
+void NaClStraceShmctl(int cageid, int shmid, int cmd,int retval) {
     fprintf(tracingOutputFile, "%d shmctl(%d, %d) = %d\n", cageid, shmid, cmd, retval);
 
 }
-void NaClStraceSocketPair(int cageid, int domain, int type, int protocol, int *fds, int *lindfds, int32_t retval) {
+void NaClStraceSocketPair(int cageid, int domain, int type, int protocol, int *fds, int *lindfds,int retval) {
     fprintf(tracingOutputFile, "%d SocketPair(%d, %d, %d, %p, %p) = %d\n", cageid, domain, type, protocol, (void *)fds, (void *)lindfds, retval);
 }
 
 void NaClStraceSecondTlsSet(int cageid, uint32_t new_value) {
     fprintf(tracingOutputFile, "%d SecondTlsSet(new_value=%u)\n", cageid, new_value);
 }
-void NaClStraceMutexCreate(int cageid, int32_t retval){
+void NaClStraceMutexCreate(int cageid,int retval){
     fprintf(tracingOutputFile, "%d mutex_create() = %d\n", cageid, retval);
 
 }
-void NaClStraceMutexLock(int cageid, int32_t mutex_handle, int32_t retval) {
+void NaClStraceMutexLock(int cageid,int mutex_handle,int retval) {
     fprintf(tracingOutputFile, "%d mutex_lock(%d) = %d\n", cageid, mutex_handle, retval);
 
 }
-void NaClStraceMutexUnLock(int cageid, int32_t mutex_handle, int32_t retval) {
+void NaClStraceMutexUnLock(int cageid,int mutex_handle,int retval) {
     fprintf(tracingOutputFile, "%d mutex_unlock(%d) = %d\n", cageid, mutex_handle, retval);
 }
-void NaClStraceMutexTrylock(int cageid, int32_t mutex_handle, int32_t retval){
+void NaClStraceMutexTrylock(int cageid,int mutex_handle,int retval){
     fprintf(tracingOutputFile, "%d mutex_trylock(%d) = %d\n", cageid, mutex_handle, retval);
 }
-void NaClStraceMutexDestroy(int cageid, int32_t mutex_handle,int32_t retval){
+void NaClStraceMutexDestroy(int cageid,int mutex_handle,int  retval){
     fprintf(tracingOutputFile, "%d mutex_destroy(%d) = %d\n", cageid, mutex_handle, retval);
 }
-void NaClStraceCondCreate(int cageid, int32_t retval){
+void NaClStraceCondCreate(int cageid,int retval){
     fprintf(tracingOutputFile, "%d cond_create() = %d\n", cageid, retval);
 }
-void NaClStraceCondWait(int cageid, int32_t cond_handle,int32_t mutex_handle,int32_t retval){
+void NaClStraceCondWait(int cageid,int cond_handle,int  mutex_handle,int  retval){
     fprintf(tracingOutputFile, "%d cond_wait(%d, %d) = %d\n", cageid, cond_handle, mutex_handle, retval);
 }
-void NaClStraceCondSignal(int cageid, int32_t cond_handle,int32_t retval){
+void NaClStraceCondSignal(int cageid,int cond_handle,int  retval){
     fprintf(tracingOutputFile, "%d cond_signal(%d) = %d\n", cageid, cond_handle, retval);
 }
-void NaClStraceCondBroadcast(int cageid, int32_t cond_handle, int32_t retval) {
+void NaClStraceCondBroadcast(int cageid,int cond_handle,int retval) {
     fprintf(tracingOutputFile, "%d CondBroadcast(cond_handle=%d, retval=%d)\n", cageid, cond_handle, retval);
 }
-void NaClStraceCondDestroy(int cageid, int32_t cond_handle,int32_t retval){
+void NaClStraceCondDestroy(int cageid,int cond_handle,int  retval){
     fprintf(tracingOutputFile, "%d cond_destroy(%d) = %d\n", cageid, cond_handle, retval);
 }
-void NaClStraceCondTimedWaitAbs(int cageid, int32_t cond_handle,int32_t mutex_handle, uintptr_t trusted_ts, int32_t retval){
+void NaClStraceCondTimedWaitAbs(int cageid,int cond_handle,int  mutex_handle, uintptr_t trusted_ts,int retval){
     fprintf(tracingOutputFile, "%d cond_timedwaitabs(%d, %d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, cond_handle, mutex_handle, trusted_ts, retval);
 }   
-void NaClStraceSemCreate(int cageid, int32_t init_value, int32_t retval) {
+void NaClStraceSemCreate(int cageid,int init_value,int retval) {
     fprintf(tracingOutputFile, "%d sem_create(%d) = %d\n", cageid, init_value, retval);
 
 }
 
-void NaClStraceSecondTlsGet(int cageid, int32_t retval) {
+void NaClStraceSecondTlsGet(int cageid,int retval) {
     // this is not used in x86 anyway
     fprintf(tracingOutputFile, "%d SecondTlsGet(%d)\n", cageid,retval);
 }
 
-void NaClStraceSemWait(int cageid, int32_t sem_handle, int ret) {
+void NaClStraceSemWait(int cageid,int sem_handle, int ret) {
     fprintf(tracingOutputFile, "%d semwait(%d) = %d\n", cageid, sem_handle, ret);
 }
 
-void NaClStraceSemTryWait(int cageid, int32_t sem_handle, int ret) {
+void NaClStraceSemTryWait(int cageid,int sem_handle, int ret) {
     fprintf(tracingOutputFile, "%d semwait(%d) = %d\n", cageid, sem_handle, ret);
 }
 
-void NaClStraceSemInit(int cageid, int32_t sem, int32_t pshared, int32_t value, int ret) {
+void NaClStraceSemInit(int cageid,int sem,int pshared,int value, int ret) {
     fprintf(tracingOutputFile, "%d seminit(%d, %d, %d) = %d\n", cageid, sem, pshared, value, ret);
 }
 
-void NaClStraceSemDestroy(int cageid, int32_t sem_handle, int ret) {
+void NaClStraceSemDestroy(int cageid,int sem_handle, int ret) {
     fprintf(tracingOutputFile, "%d semdestroy(%d) = %d\n", cageid, sem_handle, ret);
 }
 
@@ -265,11 +265,11 @@ void NaClStraceSemTimedWait(int cageid, uint32_t sem, uintptr_t trusted_abs, int
     fprintf(tracingOutputFile, "%d semTimedWait(%d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, sem, trusted_abs, ret);    
 }
 
-void NaClStraceSemPost(int cageid, int32_t sem_handle, int ret) {
+void NaClStraceSemPost(int cageid,int sem_handle, int ret) {
     fprintf(tracingOutputFile, "%d sempost(%d) = %d\n", cageid, sem_handle, ret);
 }
 
-void NaClStraceSemGetValue(int cageid, int32_t sem_handle, int ret) {
+void NaClStraceSemGetValue(int cageid,int sem_handle, int ret) {
     fprintf(tracingOutputFile, "%d semgetvalue(%d) = %d\n", cageid, sem_handle, ret);
 }
 
@@ -345,13 +345,13 @@ void NaClStraceWaitpid(int cageid, int pid, uint32_t *stat_loc, int options, int
     );
 }
 
-void NaClStraceGethostname(int cageid, char *name, size_t len, int ret) {
+void NaClStraceGethostname(int cageid, char *name, int  len, int ret) {
     fprintf(tracingOutputFile, "%d gethostname(%s, %lu) = %d\n",
      cageid, name, len, ret
     );
 }
 
-void NaClStraceGetifaddrs(int cageid, char *buf, size_t len, int ret) {
+void NaClStraceGetifaddrs(int cageid, char *buf, int  len, int ret) {
     fprintf(tracingOutputFile, "%d getifaddrs(%s, %lu) = %d\n",
      cageid, buf, len, ret
     );
@@ -363,24 +363,24 @@ void NaClStraceSocket(int cageid, int domain, int type, int protocol, int ret) {
     );
 }
 
-void NaClStraceSend(int cageid, int sockfd, size_t len, int flags, const void *buf, int ret) {
+void NaClStraceSend(int cageid, int sockfd, int  len, int flags, const void *buf, int ret) {
     fprintf(tracingOutputFile, "%d send(%d, %ld, %d, 0x%08"NACL_PRIxPTR") = %d\n",
      cageid, sockfd, len, flags, (uintptr_t) buf, ret
     );
 }
 
-void NaClStraceSendto(int cageid, int sockfd, const void *buf, size_t len,
+void NaClStraceSendto(int cageid, int sockfd, const void *buf, int  len,
     int flags, uintptr_t dest_addr, socklen_t addrlen, int ret) {
         fprintf(tracingOutputFile, "%d sendto(%d, 0x%08"NACL_PRIxPTR", %ld, %d, 0x%08"NACL_PRIxPTR", %d) = %d\n",
          cageid, sockfd, (uintptr_t) buf, len, flags, dest_addr, addrlen, ret
         );
     }
 
-void NaClStraceRecv(int cageid, int sockfd, size_t len, int flags, void *buf, int ret) {
+void NaClStraceRecv(int cageid, int sockfd, int  len, int flags, void *buf, int ret) {
     fprintf(tracingOutputFile, "%d recv(%d, %ld, %d, 0x%08"NACL_PRIxPTR") = %d\n", cageid, sockfd, len, flags, (uintptr_t)buf, ret);
 }
 
-void NaClStraceRecvfrom(int cageid, int sockfd, void *buf, size_t len, int flags,
+void NaClStraceRecvfrom(int cageid, int sockfd, void *buf, int  len, int flags,
     uintptr_t src_addr, socklen_t *addrlen, int ret) {
         fprintf(tracingOutputFile, "%d recvfrom(%d, %p"NACL_PRIxPTR", %ld, %d, 0x%08"NACL_PRIxPTR", 0x%08"NACL_PRIxPTR") = %d\n", cageid, sockfd, buf, len, flags, src_addr,
             (uintptr_t)addrlen, ret);
@@ -495,25 +495,25 @@ void NaClStraceThreadNice(int cageid, int nice, int retval) {
     fprintf(tracingOutputFile, "%d ThreadNice(nice=%d) = %d\n", cageid, nice, retval);
 }
 
-void NaClStraceNameService(int cageid, int32_t *desc_addr, int32_t retval) {
+void NaClStraceNameService(int cageid,int *desc_addr,int retval) {
     fprintf(tracingOutputFile, "%d NameService(desc_addr=%p) = %d\n", 
             cageid, desc_addr, retval);
 }
-void NaClStraceCommonAddrRangeContainsExecutablePages(int cageid, uintptr_t usraddr, size_t length) {
+void NaClStraceCommonAddrRangeContainsExecutablePages(int cageid, uintptr_t usraddr, int  length) {
     fprintf(tracingOutputFile, "%d CommonAddrRangeContainsExecutablePages(usraddr=0x%08"NACL_PRIxPTR", length=%zu)\n", cageid, usraddr, length);
 }
-void NaClStraceCommonAddrRangeInAllowedDynamicCodeSpace(int cageid, uintptr_t usraddr, size_t length) {
+void NaClStraceCommonAddrRangeInAllowedDynamicCodeSpace(int cageid, uintptr_t usraddr, int  length) {
     fprintf(tracingOutputFile, "%d CommonAddrRangeInAllowedDynamicCodeSpace(usraddr=0x%08"NACL_PRIxPTR", length=%zu)\n", cageid, usraddr, length);
 }
-void NaClStraceMmapIntern(int cageid, void *start, size_t length, int prot, int flags, int d, nacl_abi_off_t offset, int32_t retval) {
+void NaClStraceMmapIntern(int cageid, void *start, int  length, int prot, int flags, int d, nacl_abi_off_t offset,int retval) {
     fprintf(tracingOutputFile, "%d MmapIntern(start=%p, length=%zu, prot=%d, flags=%d, d=%d, offset=%lld) = %d\n", 
             cageid, start, length, prot, flags, d, (long long)offset, retval);
 }
-void NaClStraceTlsInit(int cageid, uint32_t thread_ptr, int32_t retval) {
+void NaClStraceTlsInit(int cageid, uint32_t thread_ptr,int retval) {
     fprintf(tracingOutputFile, "%d TlsInit(thread_ptr=0x%08"NACL_PRIx32") = %d\n", 
             cageid, thread_ptr, retval);
 }
-void NaClStraceThreadCreate(int cageid, void *prog_ctr, uint32_t stack_ptr, uint32_t thread_ptr, uint32_t second_thread_ptr, int32_t retval) {
+void NaClStraceThreadCreate(int cageid, void *prog_ctr, uint32_t stack_ptr, uint32_t thread_ptr, uint32_t second_thread_ptr,int retval) {
     fprintf(tracingOutputFile, "%d ThreadCreate(prog_ctr=%p, stack_ptr=0x%08"NACL_PRIx32", thread_ptr=0x%08"NACL_PRIx32", second_thread_ptr=0x%08"NACL_PRIx32") = %d\n", 
             cageid, prog_ctr, stack_ptr, thread_ptr, second_thread_ptr, retval);
 }
